Do not render buttons on single product page if there is not subscription plan
