+++ modules/ppcp-button/resources/js/modules/ActionHandler/SingleProductActionHandler.js
@@ -21,25 +21,7 @@ class SingleProductActionHandler {
         this.cartHelper = null;
     }
 
-    getPlanIdFromVariation = (variation) => {
-        let subscription_plan = '';
-        PayPalCommerceGateway.variable_paypal_subscription_variations.forEach((element) => {
-            let obj =  {};
-            variation.forEach(({name, value}) => {
-                Object.assign(obj, {[name.replace('attribute_', '')]: value});
-            })
-
-            if(JSON.stringify(obj) === JSON.stringify(element.attributes) && element.subscription_plan !== '') {
-                subscription_plan = element.subscription_plan;
-            }
-        });
-
-        return subscription_plan;
-    }
-
-    subscriptionsConfiguration(selected_variation = null) {
-        const subscription_plan = selected_variation !== null ? this.getPlanIdFromVariation(selected_variation) : this.config.subscription_plan_id
-
+    subscriptionsConfiguration(subscription_plan) {
         return {
             createSubscription: (data, actions) => {
                 return actions.subscription.create({
